QH-76 Refactor
